daemon/requirements.txt
slack-sdk>=3.27.0
schedule>=1.2.0
python-dotenv>=1.0.0
